service-bus-topics/requirements.txt
azure-core==1.30.1
azure-servicebus==7.12.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
isodate==0.6.1
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
typing_extensions==4.10.0
urllib3==2.2.1
